Move download loop to the DownloadMgr

+++ tests/test_download.py
@@ -4,9 +4,9 @@ import time
 import pytest
 import responses
 
-from prusa.connect.printer.download import Download
 from prusa.connect.printer import const
 from .test_printer import printer
+from func_timeout import func_timeout, FunctionTimedOut  # type: ignore
 
 assert printer
 
@@ -14,20 +14,6 @@ assert printer
 # pylint: disable=redefined-outer-name
 
 GCODE_URL = "http://prusaprinters.org/my_example.gcode"
-
-
-class DownloadMock:
-    def __init__(self, counter):
-        self.counter = counter
-
-    def __call__(self):
-        self.counter -= 1
-        return self.counter == 0
-
-    @staticmethod
-    def patch(loops, buffer_size=1024):
-        Download._stop_requested = DownloadMock(loops)
-        Download.BufferSize = buffer_size
 
 
 @responses.activate
@@ -45,21 +31,28 @@ def gcode(printer):
 def download_mgr(printer):
     yield printer.download_mgr
     if printer.download_mgr.current:
-        os.remove(printer.download_mgr.current.filename)
+        if os.path.exists(printer.download_mgr.current.filename):
+            os.remove(printer.download_mgr.current.filename)
 
 
-def test_download_ok_3_iterations(download_mgr, gcode):
+def run_loop(download_mgr, timeout=0.5):
+    try:
+        func_timeout(timeout, download_mgr.loop)
+    except FunctionTimedOut:
+        pass
+
+
+def test_download_ok(download_mgr, gcode):
     assert download_mgr.current is None
 
-    DownloadMock.patch(3, buffer_size=2)
     dl = download_mgr.start(GCODE_URL)
+    run_loop(download_mgr)
 
     assert dl.progress > 0
     assert dl.filename == "./my_example.gcode"
     assert dl.to_print is False
     assert dl.to_select is False
     assert dl.start_ts <= time.time()
-    assert dl.end_ts is None
     assert dl.downloaded > 0
 
 
@@ -70,21 +63,22 @@ def test_download_to_print(gcode, download_mgr):
 
 def test_download_to_select(gcode, download_mgr):
     dl = download_mgr.start(GCODE_URL, to_select=True)
-
     assert dl.to_select is True
 
 
 def test_download_time_remaining(gcode, download_mgr):
-    DownloadMock.patch(3, buffer_size=1)
     dl = download_mgr.start(GCODE_URL)
+    dl._test_loops = 1
+    dl.BufferSize = 1
+    run_loop(download_mgr)
 
     assert dl.time_remaining() > 0
 
 
 def test_download_stop(gcode, download_mgr):
-    DownloadMock.patch(3, buffer_size=1)
-
     dl = download_mgr.start(GCODE_URL)
+    dl._test_loops = 1
+    dl.BufferSize = 1
     dl.stop()
 
     assert dl.end_ts is None
@@ -93,10 +87,12 @@ def test_download_stop(gcode, download_mgr):
 
 
 def test_download_info(gcode, download_mgr):
-    DownloadMock.patch(3, buffer_size=16)
     dl = download_mgr.start(GCODE_URL, to_select=True)
-    info = dl.to_dict()
+    dl._test_loops = 1
+    dl.BufferSize = 1
+    run_loop(download_mgr)
 
+    info = dl.to_dict()
     assert info['filename'] == "./my_example.gcode"
     assert info['downloaded'] > 0
     assert info['start'] <= time.time()
@@ -114,6 +110,7 @@ def test_download_from_connect_server_has_token(printer):
     url = printer.server + "/path/here"
     responses.add(responses.GET, url, status=200)
     dl = printer.download_mgr.start(url, to_select=True)
+    run_loop(printer.download_mgr)
     assert dl.token
 
 
@@ -122,13 +119,15 @@ def test_download_no_token(printer):
     url = "http://somewhere.else/path"
     responses.add(responses.GET, url, status=200)
     dl = printer.download_mgr.start(url, to_select=True)
+    run_loop(printer.download_mgr)
     assert not dl.token
 
 
 def test_telemetry_sends_download_info(printer, gcode, download_mgr):
-    DownloadMock.patch(3, buffer_size=16)
     dl = download_mgr.start(GCODE_URL, to_print=True)
-    download_mgr.current = dl
+    dl._test_loops = 1
+    dl.BufferSize = 1
+    run_loop(printer.download_mgr, timeout=.5)
 
     printer.telemetry(const.State.READY)
     item = printer.queue.get_nowait()
